Extract capture MIDI code into its own file
