+++ src/capture/capture_midi.h
@@ -1,0 +1,26 @@
+/*
+ *  Copyright (C) 2023  The DOSBox Team
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along
+ *  with this program; if not, write to the Free Software Foundation, Inc.,
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+
+#ifndef DOSBOX_CAPTURE_MIDI_H
+#define DOSBOX_CAPTURE_MIDI_H
+
+void capture_add_midi(const bool sysex, const size_t len, const uint8_t* data);
+
+void handle_midi_event(bool pressed);
+
+#endif
