+++ vs/dosbox.vcxproj
@@ -511,6 +511,7 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClCompile Include="..\src\main.cpp" />
     <ClCompile Include="..\src\capture\capture.cpp" />
     <ClCompile Include="..\src\capture\capture_image.cpp" />
+    <ClCompile Include="..\src\capture\capture_midi.cpp" />
     <ClCompile Include="..\src\capture\capture_video.cpp" />
     <ClCompile Include="..\src\cpu\callback.cpp" />
     <ClCompile Include="..\src\cpu\core_dynrec.cpp" />
@@ -788,6 +789,7 @@ IF %ERRORLEVEL% LSS 8 SET ERRORLEVEL = 0</Command>
     <ClInclude Include="..\include\video.h" />
     <ClInclude Include="..\src\capture\capture.h" />
     <ClInclude Include="..\src\capture\capture_image.h" />
+    <ClInclude Include="..\src\capture\capture_midi.h" />
     <ClInclude Include="..\src\capture\capture_video.h" />
     <ClInclude Include="..\src\cpu\core_dynrec\decoder.h" />
     <ClInclude Include="..\src\cpu\core_dynrec\decoder_basic.h" />
