+++ vs/dosbox.vcxproj.filters
@@ -7,6 +7,9 @@
     <ClCompile Include="..\src\capture\capture_image.cpp">
       <Filter>src\capture</Filter>
     </ClCompile>
+    <ClCompile Include="..\src\capture\capture_midi.cpp">
+      <Filter>src\capture</Filter>
+    </ClCompile>
     <ClCompile Include="..\src\capture\capture_video.cpp">
       <Filter>src\capture</Filter>
     </ClCompile>
@@ -961,6 +964,9 @@
       <Filter>src\capture</Filter>
     </ClInclude>
     <ClInclude Include="..\src\capture\capture_image.h">
+      <Filter>src\capture</Filter>
+    </ClInclude>
+    <ClInclude Include="..\src\capture\capture_midi.h">
       <Filter>src\capture</Filter>
     </ClInclude>
     <ClInclude Include="..\src\capture\capture_video.h">
